requirements.txt
requests
kafka-python
psycopg2